Filter mixed-key object state safely
